Refactor the packages
The packages has been refactored
