+++ app/src/main/java/org/pesmypetcare/mypetcare/activities/MainActivity.java
@@ -37,19 +37,19 @@ import com.google.android.material.navigation.NavigationView;
 import com.google.firebase.auth.FirebaseAuth;
 
 import org.pesmypetcare.mypetcare.R;
-import org.pesmypetcare.mypetcare.activities.communication.CalendarCommunication;
-import org.pesmypetcare.mypetcare.activities.communication.InfoPetCommunication;
-import org.pesmypetcare.mypetcare.activities.communication.MyPetsComunication;
-import org.pesmypetcare.mypetcare.activities.communication.NewPasswordInterface;
-import org.pesmypetcare.mypetcare.activities.communication.RegisterPetCommunication;
-import org.pesmypetcare.mypetcare.activities.communication.SettingsCommunication;
-import org.pesmypetcare.mypetcare.activities.fragments.CalendarFragment;
-import org.pesmypetcare.mypetcare.activities.fragments.ImageZoomFragment;
-import org.pesmypetcare.mypetcare.activities.fragments.InfoPetFragment;
-import org.pesmypetcare.mypetcare.activities.fragments.MyPetsFragment;
+import org.pesmypetcare.mypetcare.activities.fragments.calendar.CalendarCommunication;
+import org.pesmypetcare.mypetcare.activities.fragments.infopet.InfoPetCommunication;
+import org.pesmypetcare.mypetcare.activities.fragments.mypets.MyPetsComunication;
+import org.pesmypetcare.mypetcare.activities.fragments.settings.NewPasswordInterface;
+import org.pesmypetcare.mypetcare.activities.fragments.registerpet.RegisterPetCommunication;
+import org.pesmypetcare.mypetcare.activities.fragments.settings.SettingsCommunication;
+import org.pesmypetcare.mypetcare.activities.fragments.calendar.CalendarFragment;
+import org.pesmypetcare.mypetcare.activities.fragments.imagezoom.ImageZoomFragment;
+import org.pesmypetcare.mypetcare.activities.fragments.infopet.InfoPetFragment;
+import org.pesmypetcare.mypetcare.activities.fragments.mypets.MyPetsFragment;
 import org.pesmypetcare.mypetcare.activities.fragments.NotImplementedFragment;
-import org.pesmypetcare.mypetcare.activities.fragments.RegisterPetFragment;
-import org.pesmypetcare.mypetcare.activities.fragments.SettingsMenuFragment;
+import org.pesmypetcare.mypetcare.activities.fragments.registerpet.RegisterPetFragment;
+import org.pesmypetcare.mypetcare.activities.fragments.settings.SettingsMenuFragment;
 import org.pesmypetcare.mypetcare.activities.views.CircularImageView;
 import org.pesmypetcare.mypetcare.controllers.ControllersFactory;
 import org.pesmypetcare.mypetcare.controllers.TrChangeMail;
