+++ app/src/main/java/org/pesmypetcare/mypetcare/activities/fragments/mypets/MyPetsComunication.java
@@ -1,4 +1,4 @@
-package org.pesmypetcare.mypetcare.activities.communication;
+package org.pesmypetcare.mypetcare.activities.fragments.mypets;
 
 import org.pesmypetcare.mypetcare.features.users.User;
