+++ app/src/main/java/org/pesmypetcare/mypetcare/activities/fragments/settings/SettingsMenuFragment.java
@@ -1,4 +1,4 @@
-package org.pesmypetcare.mypetcare.activities.fragments;
+package org.pesmypetcare.mypetcare.activities.fragments.settings;
 
 import android.app.Activity;
 import android.app.AlertDialog;
@@ -26,8 +26,6 @@ import com.google.firebase.auth.FirebaseUser;
 import org.pesmypetcare.mypetcare.R;
 import org.pesmypetcare.mypetcare.activities.LoginActivity;
 import org.pesmypetcare.mypetcare.activities.MainActivity;
-import org.pesmypetcare.mypetcare.activities.communication.NewPasswordInterface;
-import org.pesmypetcare.mypetcare.activities.communication.SettingsCommunication;
 import org.pesmypetcare.mypetcare.databinding.FragmentSettingsMenuBinding;
 import org.pesmypetcare.mypetcare.features.users.NotValidUserException;
 import org.pesmypetcare.mypetcare.features.users.User;
