+++ app/src/main/res/layout/fragment_image_zoom.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".activities.fragments.ImageZoomFragment">
+    tools:context=".activities.fragments.imagezoom.ImageZoomFragment">
 
     <androidx.coordinatorlayout.widget.CoordinatorLayout
         android:layout_width="match_parent"
