+++ app/src/main/res/layout/fragment_log_in.xml
@@ -9,7 +9,7 @@
     android:layout_marginRight="10dp"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".activities.fragments.SignUpFragment">
+    tools:context=".activities.fragments.login.SignUpFragment">
 
     <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/loginEmailLayout"
