+++ app/src/main/res/layout/fragment_new_password.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.fragments.NewPasswordFragment">
+    tools:context=".activities.fragments.settings.NewPasswordFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
