+++ app/src/main/res/layout/fragment_settings_menu.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.fragments.SettingsMenuFragment"
+    tools:context=".activities.fragments.settings.SettingsMenuFragment"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <LinearLayout
